+++ controlplane/src/core/openai-graphql/index.ts
@@ -13,31 +13,36 @@ export class OpenAIGraphql {
   }
 
   public async generateReadme(input: { graphName: string; sdl: string }): Promise<{ readme: string }> {
-    const response = await this.client.chat.completions.create({
-      model: 'gpt-3.5-turbo-1106',
-      messages: [
-        {
-          role: 'system',
-          content: `You are a friendly assistant that helps to create documentation for users based on GraphQL schema. The output must be markdown. Concentrate what data can be retrieved from it and what use cases it enables from a consumer perspective. The name of the graph is "${input.graphName}". Start with a short but easy to understand summary, conclude with key features and use cases. Add exactly one example for a valid graphql query according to the provided graph schema.
+    const response = await this.client.chat.completions.create(
+      {
+        model: 'gpt-3.5-turbo-1106',
+        messages: [
+          {
+            role: 'system',
+            content: `You are a friendly assistant that helps to create documentation for users based on GraphQL schema. The output must be markdown. Concentrate what data can be retrieved from it and what use cases it enables from a consumer perspective. The name of the graph is "${input.graphName}". Start with a short but easy to understand summary, conclude with key features and use cases. Add exactly one example for a valid graphql query according to the provided graph schema.
 `,
-        },
-        {
-          role: 'user',
-          content: `Use the following graphql schema to generate the readme:
+          },
+          {
+            role: 'user',
+            content: `Use the following graphql schema to generate the readme:
           \`\`\`graphql
           ${input.sdl}
           \`\`\`
           `,
-        },
-      ],
-      temperature: 0,
-      max_tokens: 1000,
-      n: 1,
-      top_p: 1,
-      frequency_penalty: 0,
-      presence_penalty: 0,
-      stream: false,
-    });
+          },
+        ],
+        temperature: 0,
+        max_tokens: 1000,
+        n: 1,
+        top_p: 1,
+        frequency_penalty: 0,
+        presence_penalty: 0,
+        stream: false,
+      },
+      {
+        timeout: 60_000,
+      },
+    );
 
     return { readme: response.choices[0].message.content ?? '' };
   }
@@ -47,18 +52,19 @@ export class OpenAIGraphql {
       sdl: z.string().describe('The fixed GraphQL Schema'),
     });
 
-    const res = await this.client.chat.completions.create({
-      model: 'gpt-3.5-turbo-1106',
-      temperature: 0,
-      top_p: 1,
-      n: 1,
-      frequency_penalty: 0,
-      presence_penalty: 0,
-      stream: false,
-      messages: [
-        {
-          role: 'user',
-          content: `Given the following SDL (Schema Definition Language) from an Apollo Federation subgraph:
+    const res = await this.client.chat.completions.create(
+      {
+        model: 'gpt-3.5-turbo-1106',
+        temperature: 0,
+        top_p: 1,
+        n: 1,
+        frequency_penalty: 0,
+        presence_penalty: 0,
+        stream: false,
+        messages: [
+          {
+            role: 'user',
+            content: `Given the following SDL (Schema Definition Language) from an Apollo Federation subgraph:
                     
                     \`\`\`graphql
                     ${input.sdl}
@@ -72,17 +78,21 @@ export class OpenAIGraphql {
                     
                     Please take into consideration the schema check result and generate a fixed version of the SDL.
                     Once completed, store the SDL using the 'store_sdl' function.`,
-        },
-      ],
-      functions: [
-        {
-          name: 'store_sdl',
-          description:
-            'This function allows the Agent to store the fixed SDL. The SDL must be returned in plain text as a GraphQL Schema.',
-          parameters: zodToJsonSchema(out),
-        },
-      ],
-    });
+          },
+        ],
+        functions: [
+          {
+            name: 'store_sdl',
+            description:
+              'This function allows the Agent to store the fixed SDL. The SDL must be returned in plain text as a GraphQL Schema.',
+            parameters: zodToJsonSchema(out),
+          },
+        ],
+      },
+      {
+        timeout: 60_000,
+      },
+    );
 
     try {
       return out.parse(JSON.parse(res.choices[0].message!.function_call!.arguments!));
