+++ studio/package.json
@@ -84,7 +84,7 @@
     "lodash": "^4.17.21",
     "monaco-editor": "^0.45.0",
     "monaco-graphql": "^1.4.0",
-    "next": "14.0.4",
+    "next": "14.1.0",
     "next-themes": "^0.2.1",
     "prism-react-renderer": "^2.0.6",
     "prismjs": "^1.29.0",
